Use defined logging levels with glog's V-style logging
Introduces a set of defined logging levels that can be
passed to glog.V() while logging based on the log level configured.
This provides some level of consistency and reasoning for the logs
instead of clients arbitrarily using values directly.

The logging levels do not have consecutive values to allow
for newer intermediary levels that could be added in the future.

resolves #69