ref(event notifications): Use a broker to relay events to xDS streams
Resolves #118
Previously a single announcement channel was used to relay announcements
from informers to xDS streaming functions. This change introduces
a broker for this purpose, and refactors code as necessary.

Each event informer uses its own announcement channel, and the broker
waits on events on these channels. When a new event is read, the broker
publishes the message over a dedicated channel for each proxy registered
with the catalog.

Testing done:
- Verified demo app: bookbuyer calls returned status code 200 (*with
  some delay)
- Verified traffic target