fix(service-cache): Skip adding empty endpoints
Adding empty endpoints to the service cache causes readers of the
cache to assume the service name is resolved to one or more endpoints.
Noticed that without this change, EDS was reporting invalid endpoints in
the DiscoveryResponse message to envoy.

The demo app would previously return 503 status code consistently, but
with this change gives a 200 status code as expected.