webhook: Prefix the name with the namespace (#310)
CI does 2 parallel builds minimum - more if there are more than one PRs in a given moment. Currently we deleting and then create a webhook wit heach PR but these always have the same name (ads). The web hooks point to different ADS. Each hook is supposed to point to the ADS in the namespace dedicated to the CI run it is created by.
So essentially we can only have one webhook pointing to one of the ADS that may be running.

This PR perfixes the name of the webhooks with the namespace of the ADS for which they are meant. So if the webhook is called ads -- I propose we prefix it with ci-552706-1794 which is the $K8S_NAMESPACE env var.