demo: Only exit on failure for bookbuyer and bookthief
It has been observed that when the bookstore POD is ready
before the bookbuyer POD, the bookbuyer HTTP GET calls to
the bookstore service might succeed too soon. While this
is expected, it causes the bookbuyer container to exit
causing a CrashLoopBackoff of the POD in K8s.
This change avoids this scenario by only exiting on failure.

Resolves #331