kube/client.go: Don't use net.ParseIP while creating endpoint (#387)
This change fixes the demo which breaks when net.ParseIP is used
to create the endpoint object.
net.ParseIP has side-effects where it can return nil if the IP
is not valid. Suspect this to be causing encoding issues while
programming Envoy, as opposed to an empty string.

The demo works as expected with this change.