deploy/xds: Use upstream Envoy image (#388)
Currently a cached image from a container registry defined
by CTR_REGISTRY env variable is used for the envoy image. This
is unintentional.

This change fixes this to use the latest envoyproxy/envoy-alpine
image. Although we should pin the version before releasing, tagging
on latest helps us quickly find incompatibilities with a released
envoy version. We already know the current released version is
v.1.13.1 on envoy-alpine, so if things break we can simply revert
to using this version.