feat(logging): Use zerolog for structured logging
This change replaces glog with zerolog as the logger
for OSM. Zerolog has many benefits over glog, with
structured logging being the main motivation to use it.

Since the logging methods are different, this change updates
all log messages to use the zerolog format, most notably for
logging errors. The log levels are also updated making it
easier to specify a human readble log level instead of a number.

Resolves #385