Fix net.IP conversion to/from string type
net.IP is a byte slice, hence a string representation of an IP
address should use net.ParseIP to create an IP address. The reason
Envoy was not being programmed correctly with this change alone
is because of incorrect conversion from net.IP to string type
while programming Envoy.

This change fixes this issue and ensures the demo works as expected
without any type conversion issues.