logging: Use component name instead of the package name
A package name might not be a good way to identify a component
logged in the "component" field of the log message. There can
be nested packages, packages with the same name etc. that makes
package name a bit ambiguous in logs. Since we log the filename
and line number, logging the package name doesn't provide much
value considering multiple packages with the same name exist
in this project.
This change uses a better readable name for components that use
subloggers. It also obviates the need to use package name in
log fields thereby allowing files within the same package to use
different subloggers if necessary.