Program HTTP methods for a route instead of virtual host
In the SMI HTTPRouterGroup spec, a domain could have multiple
HTTP method access policies and hence multiple associated routes.
Currently OSM programs the allowed methods on the top level
virtual host instead of the specific routes within a virtual host.
This will not work when #415 is fixed, where a virtual host will
be associated with a domain. An access policy for a route takes
precedence over the virtual host policy, so program the allowed
method associated with a route as a part of the route matching
for a route within a virtual host.

Also updates the traffic spec to include multiple HTTP methods
for sanity purpose and the import identifier for envoy's route
package to avoid conflicts.

Resolves #455