catalog: Issue certificates for services based on GetCommonName (not the service name) (#484)
When the service mesh catalog requests for a new certificate to be issued, it needs to come up with the CN for the certificate. The CN is currently generated based on the namespace and the name of the service for which the certificate is issued. The namespace/service is not a correct DNS-1123 FQDN. For that reason we added GetCommonName (via #486) and in this PR we are using it to issue a correct certificate.

Without this correction certain certificate issuers (such as Hashi Vault) would not be able to issue certs (error out w/ incorrect CN).

This is one of the many PRs in the Vault integration series (#426).