envoy/route: Methods for a route should be separate routes
If a route for a path regex /books-bought has multiple allowed
HTTP methods such as GET, POST, then the header match for the
route should contain individual methods in separate routes
instead of being a part of the same route.
Different HTTP method header matches cannot be a part of the same
route because all headers are matched for a route configuration
to be selected and its not possible to have an HTTP request with
multiple methods set.

This change decomposes HTTP methods for a route path into separate
routes.

Also updates and imrproves the unit test by removing hardcoded
expectations for values.

Resolves #508