Map a proxy instance to a service instead of service account (#516)
This change (#515) is part of a broader issue (#514) to constrain a service
running in a mesh to a service account. It encodes the service name
in the proxy's certificate instead of its service account. The allows
to remove the hack to pick the first service in the service account
while creating the proxy instance in ADS.

In order to map 1 proxy -- 1 service -- 1 service account, this change
does the following:
1. Uses openservicemesh.io/osm-service annotation as an explicit way
   to infer a service running on the POD
2. If the annotation is not present, assumes the service and service
   account have the same name.

With 1., service and service accounts can have different names as in
the demo setup.
With 2., service and service accounts should have the same name.

This change does not constrain 1 service -- 1 service account mapping,
which would be the responsibility of pre-OSM install checks or
validation webhooks.