injector: Move from a multiline string YAML template to Go struct to YAML function (#527)
This PR introduces the function getEnvoyConfigYAML(), which generates the same thing as what is currently contained in envoyBootstrapConfigTmpl (as verified by a unit test).

This gives us a bit more compile time certainty that the Envoy config is to some degree correct. (Ensures YAML is valid, properly structured/shaped etc)

In this PR we deliberately keep envoyBootstrapConfigTmpl where it is (instead of moving it to the tests, since now it is only serving the purpose of validating that go-yaml correctly converted the struct to YAML) in order to show that go-yaml generates the exact same YAML. In another PR this will be moved to tests.