ingress: catalog changes for ingress routes and endpoints
This change adds support for mesh catalog to consume ingress
resources from ingress monitor and convert them to route
policies per domain. It also adds an API in catalog to list
endpoints for a service, without having a dependency on the
service to be referenced in SMI. Ingress endpoints will not
have an SMI reference but will need to resolve to K8s provider
endpoints.

Also adds unit tets for the ingress route policies.