demo: fix traffic split usage
The demo has been interpreting SMI traffic split incorrectly.
Per the SMI traffic split policy, `spec.service` should be a
valid k8s service, and so should the leaf services specified
by `spec.backends[].service`. SMI does not impose any constraint
on what the leaf services should be other than the fact that
they belong to the same namespace as the root service in the Split.

The reason the root service `spec.service` is required to be a
valid k8s service is because `spec.service` is the FQDN that
apps use to communicate. This FQDN needs to be resolvable by
k8s, which would not be possible if `spec.service` is not a
k8s resource.

The demo has been using `bookstore.mesh` as the FQDN for the
bookstore service which is incorrect, but this has been working
because of a hack to statically resolve it to a random IP. Per
SMI (160), the FQDN should correspond to the root service in the
Split definition, obviating the need to statically configure
DNS in client pods.

This change fixes the demo to adhere to the SMI Split policy as
expected. One should note that since the demo deploys apps in
different namespaces, DNS queries for apps in different namespaces
should be of the form <service>.<namespace>. The demo uses
bookstore-v1 as the root bookstore service, with traffic split
between bookstore-v1 and bookstore-v2.