catalog: Adding Expected proxy (#583)
At some point I remember discussing the idea of accounting for certificates issued to Envoy proxies for XDS connections, and then comparing these to the actual proxies connected.

This would be extremely useful as part of the support tools -- the issued certificates minus connected proxies gives us the proxies that may have run into some trouble.

This not a finished product just yet. I imagine we'll have a way to query for this data from the outside world (perhaps as part of our logging/metrics work).