Make k8s event eventually consistent
Updates for k8s events were ignored if the old and new
objects are the same. This breaks the eventual consistency model
where an event could be missed/lost by the system due to network
faults, consumers being down etc. This change processes k8s update
events for monitored namespaces based on the default resync interval.
The point of even having resync interval for events is to achieve
eventual consistency.