namespace: Populate cache of NamespaceService w/ changes to k8s namespaces via k8s events (#631)
We saw intermittent demo/ci failures. It seems to me that these may be related to inconsistent Namespaces cache. The namespace controller needs informer.AddEventHandler added to it so that if the namespaces were tagged after this was started it will pick up the events and register the namespaces as being observed.

I discovered this by looking at ADS pod logs in one of the CI namespaces - noticed that ADS reported the bookstore/bookbuyer namespaces as "Not monitored: xxx", which is a reason for demo to not work.

What's up with the nsFilter function?
We need to add AddEventHandler from our pkg/kubernetes into pkg/namespace, but then also pkg/kubernetes needs a function from pkg/namespace - so we endup with a cyclical import ∞.

To move this forward and fix CI I decided to create a generic filter function (could have been a variadic functions...). The NamespaceController needs AddEventHandler, but it should NOT have a filter (because it will filter the namespace events, which it needs to watch to understand whether they are tagged to start watching the namespace events ↻)

So we make a type namespaceFilter func(obj interface{}) bool and let each Controller that needs to manage cache define their own filter. NamespaceController should have no filters - so it passes nil. Everybody else makes the same namespace filter function.