Add a readiness probe for the injector webhook
It has been observed that in some cases the sidecar injection
webhook is not invoked for pods being deployed around the same
time as ADS. To rule out such race conditions in the demo, add
a readiness probe for the webhook so that we are guaranteed that
the webhook is ready by the time ADS pod is marked as ready by K8s.
The demo ensures the ADS pod is ready before deploying the apps.

Could potentially fix #627