catalog: fix pod lookup for proxy certificate
There is a bug where the pod being matched to a proxy
certificate is taking the address of the loop variable,
which is the same for each iteration of the loop. While the
pointer is the same in each iteration, the object being
pointed to is overwritten in each iteration, causing
the pointer to hold the value of the last iteration.

This was basically resulting in a matched pod being
overwritten with the value of the loop variable
from the last iteration.

Resolves #633