ci: Relax the ci from flagging failures after a previous success
Commit 931056d5122f0e4a201bcb7a34e68bb63b1b1d05 hardened the CI
in a few ways. This change relaxes the CI to avoid flagging CI
runs where a failure is seen after a previously seen success,
because this continues to happen in the CI intermittently.
Debugging why a particular request failed in the CI is going to be
hard just looking at the logs.