Moving CI_MIN_SUCCESS_THRESHOLD from github secrets to in-repo workflow/main.yaml (#757)
Last week we added CI_MIN_SUCCESS_THRESHOLD as a Github secret. This is/was set to 1.
Because it is in a Github secret - every time we see a 1 in stdout of the CI system - it gets rewritten to ***

It should not be a problem to move this variable inside .github/workflows/main.yml and tweak it from there by submitting PRs.