Refactor SDS (#760)
In the process of troubleshooting #661 I realized the SDS code is overly complicated. (I take responsibility - wanted to use functional approach and dry-it up - but it became tough to understand)

This PR refactors the SDS code to make it a) shorter b) easier to understand

This does not address #661 but I believe this is a step forward to understand.