Remove CI_MAX_WAIT_FOR_POD_TIME_SECONDS from Github secrets; hardcode in github workflow (#758)
The env var CI_MAX_WAIT_FOR_POD_TIME_SECONDS is/was added to Github secrets. The value set to 30.

Anytime github sees 30 in stdout of CI - it converted it to *** because it is assumed to be a secret. This is giving us/me hard time reading logs (similar to #757)

Moving the value of CI_MAX_WAIT_FOR_POD_TIME_SECONDS from Github secrets to the workflow committed to the repo solves this.