Set cmd line bool options correctly
The golang flags package works different with bool options.
A bool option does not require a value of true/false. The
presence of the option is sufficient to determine if the
option is set. Moreover, the `flags` package expects values
for bool options to be passed using `-option=value` as opposed
to `--option=value`. This results in confusion and is prone to
bugs, so use `--option` without a value to signify the setting
of a boolean option.

This change adds a new boolean option to the cli to disable
SMI access control policy and updates the debug server option
to be set correctly.