ci: use a separate timeout for success
This change uses the existing CI_WAIT_FOR_OK_SECONDS env variable
to configure the time the CI should wait for a success. Previously
the pod wait time was used to time out the success test as well.
This allows fine tuning of the pod wait time and success wait time.

Additionally the success wait time is increased to 60s to give the
CI more of a chance to succeed rather than timing out before a
potential success, which would cost more time due to having to
rerun the entire CI build again.