debug: Provide a way for OSM pod to port forward from a side car to OSM (#800)
The Envoy proxies in Open Service Mesh expose an admin port 15000, which has a few endpoints allowing us to get Envoy's current config, certificates etc.

Typically we troubleshoot service mesh issues by using kubectl port-forward and then browsing localhost:15000

It would be great if the OSM pod itself had access to that information for each Envoy proxy. This would be very useful to compare state of what OSM sent to the proxy and what the proxy's current config is.

For instance -- it would be terrific to be able to see what certificates are in OSM Pod's cache and what certs are currently on an Envoy. (Compare on cert's serial number for instance.)

It is not possible for the OSM pod to directly query the Envoy pods on 15000, but it is possible to port forward that to the OSM pod.

This PR adds the ability to port forward from a given Envoy to the OSM pod and introspect Envoy's config. This is an experimental feature. It is very helpful, but we should consider eventually moving it into a different pod. This is only enabled when the -enable-debug-server flag is on (off by default)