ci/maestro: Wait for all service pods to be ready before test
Currently maestro only waits on bookbuyer and bookthief pods
to be ready before starting to test if HTTP responses are
as expected. The requests will fail if bookstore (and bookwarehouse
eventually) are not ready. This change makes maestro wait on
all service pods to be ready before starting the timer for
the test to succeed.