certificate: fix certificate rotation updates
Despite certificates being rotated, the rotated certificates
were never being consumed by catalog. This is because catalog
maintained its own cache for certificates while rotated certs
were a part of CertManager's cache.

This change resolves this issue by solely relying on CertManager's
cache as the source for cached certificates. If a certificate
is not found in the cache, catalog will request CertManager to
issue a new certificate.

Resolves #661