Add Inbound and Outbound to MatchSubjectAltNames (#766)
With #747 we added a bookWarehouse. The bookstore gets inventory from the bookwarehouse.

Because of mTLS - when bookstore connects to bookwarehouse

bookstore (the client) verifies that the warehouse certificate is signed with the root cert and is in the list of MatchSubjectAltNames
bookwarehouse (the server) verifies that the bookstore cert is signed by the root cert and is in the MatchSubjectAltNames
We populate MatchSubjectAltNames using the ListAllowedPeerServices function, which lists services per destination only.

The connection between bookstore and bookwarehouse is currently not working. We see the following error we see in the bookstore Envoy logs:
[2020-06-04 23:04:21.291][14][debug][connection] [source/extensions/transport_sockets/tls/ssl_socket.cc:226] [C204] TLS error: 268435581:SSL routines:OPENSSL_internal:CERTIFICATE_VERIFY_FAILED

See the following PR, which introduces a check of bookwarehouse in CI: #767
The PR above will start passing once this here is merged.

The SSL verification error s is caused by the MatchSubjectAltNames on the bookstore pod, which fails to validate the certificate of the bookwarehouse. We have a function ListAllowedPeerServices which lists all permitted inbound certs. Inbound means: on bookstore we allow bookbuyer. On warehouse we allow bookstore. But we also need to allow warehouse on bookbuyer for mTLS (m for Mutual, where bookstore checks warehouse eventhough bookstore is the client). In other words ListAllowedPeerServices does not list the outbound.

To solve this:

renamed ListAllowedPeerServices to ListAllowedInboundServices
created ListAllowedOutboundServices