cds: Simplify local cluster config
This change always creates a local cluster for a service
to handle incoming traffic. The local cluster is used
for both in-mesh and ingress traffic to the service.
Even if a local cluster is created, only the presence of
a route ensures that the local cluster can be reached, so
there should be no downside to creating a local cluster
at all times. This simplifies the code and voids the need
to create a local cluster separately for ingress.