ingress: remove monitor annotation
The monitor annotation doesn't add much value, considering
the ingress resource needs to belong to a monitored namespace.
When a user applies an ingress resource with backend services
that belong to the service mesh, the intent is clear that the
the service needs to be exposed outside the cluster via ingress.
OSM doesn't use the monitor annotation to match a service against
an ingress resource, but rather the service name and the namespace
of the ingress resource.