Make the sidecar injector webhook's failurePolicy configurable
By default, the failurePolicy is set to Fail to prevent pods
from being injected into the mesh if the sidecar injector webhook
fails. This is the desired behavior in most cases so as to get
instant feedback when a sidecar needs to be injected but fails.
However, this setting is not suitable in all environments, AKS
being one of them where the requirement is for the failurePolicy
to be set to Ignore due to limitations in how AKS handles webhook
failures.

This change makes the webhook's failurePolicy configurable via the
install cli and Helm chart.

Resolves #908