Make the sidecar injector webhook's failurePolicy configurable
By default, the failurePolicy is set to Fail to prevent pods
from being injected into the mesh if the sidecar injector webhook
fails. This is the desired behavior in most cases so as to get
instant feedback when a sidecar needs to be injected but fails.
However, it might be desirable in some environments to make the
failurePolicy configurable.

This change makes the webhook's failurePolicy configurable via the
install cli and Helm chart.

Resolves #908