fix(demo): skip logging success message after timeout (#920)
Previously, even after the demo timed out, a success message would be
printed. Since a failure message is always printed after the timeout,
the same iteration would indicate both success and failure if the
expected return code is returned for a request past the deadline. By
default, no deadline is defined in the CI_MAX_ITERATIONS_THRESHOLD
environment variable, so success messages will be logged
forever. For CI, this variable is overwritten to its previous default
value of 30.