Revert "Created a CRD and Go types for backpressure in OSM (#850)"
This reverts commit f4d20579893d5946a811247f877999b41fa7df5d.

The plan is to develop the backpressure feature in a separate
branch called `experimental`. The go client and API autogen
code is going to change as we iterate over the CRD. The feature
will also require the CRD to be installed along with modifications
to OSM RBAC rules which affects mainline features.

Once we have settled on v1alpha1 of the spec, experimental
changes will be merged into master again.