Revert "Make the sidecar injector webhook's failurePolicy configurable (#939)"
This reverts commit 489f98d5ac8a16ccf3168137ee8c28dd60891493.

Allowing the failurePolicy to be set to Ignore opens the door
for the mesh state to be inconsistent from the desired state
when traffic should be denied by default. If the webhook fails,
then sidecars will not be injected and without the sidecar
traffic will not be intercepted for policy enforcement.