Test fixes
xdsutil_test.go:
With updated packages:

go vet

The fact it's not complaining now is just subject to the version of packages/go
we are using. This is bound to happen at some moment.
Changing to compare the pointers instead will properly compare the objects
without making a copy of the inner state of the wrappers (which hold a mutex).
(And no, the addresses of the pointers are not compared)

response_test.go:
Re-adding comparing the value of the protobuf. We will manually rebuild it
for the time being.