configurator: Implement ConfigMap informer (#1029)
This PR implements a ConfigMap informer. This will start as soon as the controller starts. It has a very limited scope - will watch osm-config in the OSM namespace.

The ConfigMap will be exposed to the rest of the OSM components in a follow up PR. So far the ConfigMap only implements allow_all key.