Adding get-pcap helper script + port forwarding scripts (#1053)
This script captures certain amount of network traffic from the BOOKSTORE V1 pod. The capture is then moved from pod to local host. This is a pcap file that can be inspected with tools like Wireshark. This is useful to show encrypted/unencrypted traffic (demo enabling mTLS).

This PR also adds helper scripts that enable port forwarding from the various demo pods. This makes it easy / possible to bring the various web UIs to local host and show the count of books.