Update envoy-go-sdk version (#1063)
Envoy version 1.14.0 marked certain definitions (such
as filter names) for deprecation. When we updated the
envoy image version in the past, the go-sdk wasn't
updated, so the newer definitions were never available
to use.

This change updates the envoy-go-sdk package version
and ensures the newer definitions that are compatible
with the current Envoy image are used.