demo: revert bookstore port to 8080 (#1102)
The bookstore service is used to demo HTTPS ingress. Azure
Application Gateway does not support using port 80 for HTTPS:
https://github.com/Azure/application-gateway-kubernetes-ingress/ \
    blob/master/docs/annotations.md#backend-protocol

PR #1009 changed the port to be able to demo brownfield scenarios.
Since the brownfield demo is a work in progress, revert this
change to avoid breaking ingress. A subsequent change will
enable brownfield scenarios to be demoed.

Resolves #1099