demo: Do not use GIT_HASH in the identity of the bookstore (#1107)
This PR removes the git hash from the identity of the bookstore. This identity is passed via HTTP response header back to the bookbuyer and bookthief and are displayed in the web UI. This answers the question "who did I purchase a book from". For better flowing demo I decided to remove this, which leaves us with bookstore-v1 and bookstore-v2 -- this instead of bookstore-v1-abcdefg