README/env: update example env file
New developers are using the example file as their base env file
to run the demo. This however doesn't work because the values
in .env.example are samples and left to the user to configure.
Since this introduces confusion as to what values should be put
in the env file, comment out optional exports of env variables
by default.

The example env file will only have the defaults required to run
and observe the state of the demo. All other optional env variables
are for documentation purpose only, and need to be uncommented
by the user based on their demo configuration.

Resolves #1117