ref(demo): allow empty $CTR_REGISTRY_PASSWORD
This change reuses the create-container-registry-creds.sh script in
configure-app-namespaces.sh instead of reimplementing it to skip
creating image pull secrets when no $CTR_REGISTRY_PASSWORD is defined.

scripts/create-container-registry-creds.sh now takes a required argument
for the namespace in which to store the secret. It also now handles the
$DOCKER_PASS variable for CI, and the CI-specific versions of the script
have been removed.