envoy: consolidate cluster creation and use SDK value
This change refactors `envoy/cds` by consolidating
cluster creation code in `cluster.go`. The functions
being moved should not be in xdsutil but rather a
part of cds.  It renames some functions to better convey
their purpose. The connect timeout for clusters is
also adjusted for consistency.

The change also used the name of the TLS transport
socket as defined in the SDK.