egress: use mesh cidr for in-mesh routing (#1182)
This change reworks how in-mesh and egress traffic is routed when
egress is enabled. When egress is disabled, in-mesh traffic gets
routed as before - all outbound traffic is routed using a single
filter chain meant for in-mesh traffic.

When egress is enabled, the mesh CIDR ranges (service and pod CIDRs)
need to be configured. The outbound routing is configured such that
all destinations belonging to the mesh CIDRs get routed to using
the in-mesh filter chain, while any traffic that does not match the
mesh CIDR (egress traffic) matches the egress filter chain. This simplifies
the outbound routing logic. The benefit is that egress is not limited to
HTTP/HTTPS traffic, but will work for any kind of egress traffic (ex. raw
TCP).
This also allows to enforce policies for in-mesh traffic at the source,
whether egress is enabled or not, resulting in consistent HTTP response
codes when an unauthorized in-mesh source attempts to send traffic to an
in-mesh destination. As a result bookthief in the demo will get a 404
response irrespective of egress.

The change also updates the CI to enable egress for one of the tests.

Part of #999, #1000, #1120