catalog: Remove IsIngressService from interface (#1195)
The goal of this PR is to remove IsIngressService and reuse the underlying function. The implementation of this is simple enough to where it does not really need to be part of the mesh catalog interface.

This PR also is the beginning of switching from using the word domain to host.