ref(demo): use --ignore-not-found for kubectl delete commands
Many `kubectl delete` commands in the demo have a `|| true` guard on the
end to prevent failure when the resource doesn't exist. This change
replaces that guard with the `--ignore-not-found` flag of `kubectl
delete`. This flag better communicates intent, allows the commands to
fail for other errors we are not able to handle, and prevents error
messages from being displayed when the resources do not exist.