fix(demo): apply proper labels to bookstore svc
* The root service should have the app=bookstore label selector
so it can direct traffic to Pods that have the same label or a
Traffic Split hasn't been applied or the services that point to
version specific bookstore pods don't exist
* Given these changes, the root service will target Bookstore pods
regardless of the version of Bookstore deployed. The root service will
match all bookstore pods where there is a label app=bookstore.