ref(tools): manage tools with go.mod (#1338)
Currently, tool dependencies like those for linting and test coverage
have to be installed manually by calling `go get` before they are
invoked. This is both error prone and subject to version skew. This
change adds those dependencies to a special tools.go file with blank
imports for those dependencies with a `// +build` constraint to avoid
them from being compiled into any binaries. Then, to invoke the tool,
instead of calling the binary and assuming it's in the current `$PATH`,
call `go run` with the full module path.

e.g. instead of

    golint -set_exit_status ./...

use

    go run golang.org/x/lint/golint -set_exit_status ./...

Even though this style of invocation is less ergonomic, tools are most
often used from scripts and tools can still be installed to `$GOBIN` to
be used manually. This change is primarily to pin the versions of tools
used for CI and eliminate ad hoc `go get` calls whenever tools are used.

More details about this technique are documented in the [Go wiki](https://github.com/golang/go/wiki/Modules#how-can-i-track-tool-dependencies-for-a-module).