Use kubernetes hostnames for root service in a traffic split
When clients make HTTP requests to the root service (spec.service)
in the SMI split policy, the requests could be directed to a specific
port. If the destination service is using ports other than 80, then
clients will make HTTP requests by specifying the port.
Ex. GET http://bookstore.default.svc.cluster.local:9090 to access
a bookstore service in the default namespace running on port 9090.

The above scenarios would result in the port being included
as a part of the host header in HTTP requests, which will not
match against the L7 HTTP rules OSM configures because OSM does not
currently program different combinations of the root service's
hostnames. This change allows clients to be able to make HTTP
requests using kubernetes hostnames for a root service. This
is similar to how hostnames are configured for permissive mode
and when no traffic split policy is specified, with the difference
being the hostnames point to the root service in a traffic split.

Resolves #1245