envoy/permissive-mode: use cluster provided service discovery and routing
When operating in permissive traffic policy mode, OSM relies on native
kubernetes networking to route traffic to services. Since Kubernetes
is already aware of how to discover and route traffic to services,
rely on this instead of the control plane to explicitly program and
discover endpoints. This change makes the controller logic simpler for
permissive mode and is an optimization over the current implementation.

Resolves #1403