envoy/rds: use route direction while creating routes
The current code relies on passing bools around for source/destination
route configs and if the route corresponds to a local cluster. This can
be confusing and hard to understand if one does not know how these
booleans are interpreted in the code.
This change simplies the logic by introducing the concept of a direction
for a route. Routes can be outbound (to reach a remote service), or
inbound (to accept traffic on the local cluster host). Since the route
configurations are alredy using Inbound/Outbound terminology, updates
the variable names to reflect this.

There is no functional change introduced in the code, and the routing
logic remains as before.